moving use fetch to server
